[0.0.3] Moving the 'printing time' command to its own package.
